+++ src/index.ts
@@ -6,6 +6,7 @@ export * from "./components/forms/Form.js";
 export * from "./components/forms/FormCheckboxInput.js";
 export * from "./components/forms/FormEmailInput.js";
 export * from "./components/forms/FormFieldset.js";
+export * from "./components/forms/FormHiddenInput.js";
 export * from "./components/forms/FormLabel.js";
 export * from "./components/forms/FormNoticeContainer.js";
 export * from "./components/forms/FormNumberInput.js";